Twitter_Demo
The project is a web application like Twitter, where users can create an account, log in, post and like. Java Spring Boot is used on the Back-end side of the project, and Html, Css, React.js is used on the Front End side, but the project is not completed yet, it is still ongoing.